+++ poi-ooxml/src/main/java/org/apache/poi/xssf/usermodel/XSSFColor.java
@@ -248,7 +248,7 @@ public class XSSFColor extends ExtendedColor {
      */
     @Override
     public int getTheme() {
-        return (int) ctColor.getTheme();
+        return Math.toIntExact(ctColor.getTheme());
     }
 
     /**
